Import from Patchwork master
This repo was generated by splitting out the history of the pwclient
script in its various iterations from the main Patchwork tree [1] at
commit 810d2c92.

This was achieved as follows:

  $ mkdir pwclient && cd pwclient && git init && cd -
  $ git clone https://github.com/getpatchwork/patchwork && cd pwclient
  $ git checkout -b pwclient-split && git checkout -
  $ PATHS_TO_KEEP="patchwork/bin/pwclient apps/patchwork/bin/pwclient apps/patchwork/bin/pwclient.py COPYING"
  $ git filter-branch -f --index-filter "git rm --ignore-unmatch \
      --cached -qr -- . && git reset -q \$GIT_COMMIT \
      -- $PATHS_TO_KEEP" --prune-empty -- "pwclient-split"

which, in brief, filters out the history for the pwclient file, in the
many locations it has been located, and COPYING file. After this, the
'pwclient' file was still nested in 'patchwork/bin' so history was
manually rewritten (via rebase) to avoid this. This dropped one commit
which probably never made sense in this split out context anyway:

  ad2762cf Move to a more recent django project structure

Finally, the branch was pushed to the new project.

[1] https://github.com/getpatchwork/patchwork

Signed-off-by: Stephen Finucane <stephen@that.guru>