action_get: Fix race condition in patch creation
An attacker might foresee that a certain patch is about to be retrieved
with

  pwclient get $id

When this is called in a globally writeable directory (e.g. /tmp) the
attacker can with careful timing create a symlink with the patch's name
(that is provided by the patchwork instance) to some precious content
just after the check for os.path.exists(fname) but before this file is
opened for writing. As a result the content is overwritten.

Fix this by using 'x' as mode for writing which ensures that the file
doesn't exist. Note however this requires Python >= 3.3.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Reviewed-by: Stephen Finucane <stephen@that.guru>