Resolve check-related TODOs
We're still preparing for the REST API implementation. Two precursor
changes are needed related to the REST API's nesting of checks
underneath their parent patch. Because of this, it is not possible to
list checks without specifying a patch. Similarly, it is not possible to
show an individual check without providing both the check ID and the
patch ID. As a result, the 'check-list' command gains a 'PATCH_ID'
positional argument as well as a '--user' option, while the 'check-info'
command now accepts a 'PATCH_ID' positional argument as well as the
'CHECK_ID' argument. These are all optional for the XML-RPC API
implementation (for now) but that will not be the case for the REST API.

Signed-off-by: Stephen Finucane <stephen@that.guru>