Add REST API proxy layer
We're finally there! This provides a REST API layer that can, for the
most part, be switched in seamlessly in place of the XML-RPC API layer.
All that remains now is to provide a way to switch between the proxy
layers. That'll require some config changes so we'll do that soon.

All this work to preserve an old, legacy API that no one cares about.
Really makes you think.

Signed-off-by: Stephen Finucane <stephen@that.guru>