Remove unnecessary check for missing args
If argparse has an optional parameter and a required parameter but only
one parameter is provided at the command line then the optional
parameter will be rightly initialised to None. We had a check that
assumed the opposite. We can remove this logic and update the help text
for the option to reflect the real world behavior.

Signed-off-by: Stephen Finucane <stephen@that.guru>