xmlrpc: Enable 'allow_none'
This is a custom extension to the XML-RPC spec so it's not enabled by
default. In truth, the server shouldn't be returning `None` values but
there's a bug somewhere. The server is frozen so we need to fix this
here, on the client side.

While we're here, 'Server' in the 'xmlrpc.client' library is a legacy
alias for 'ServerProxy'. Update the call.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Closes: #22