Don't expect return value from API.patch_set
The XML-RPC API returns True if a patch update is successful else it
raises an exception. The REST API returns an updated copy of the patch
if successful else it returns a HTTP 4xx error. These different return
styles meant we couldn't effectively return a value from the 'patch_set'
API method in the success case, only raise an exception in the failure
case. We addressed this in commit 0949e9e but forgot to update callers
than were checking for a success return value. Do this now.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Fixes: 0949e9e ("api: More signature corrections")
Closes: #23