convert REST API to use requests library
Switch the REST API implementation over to the requests python library.
This will make it easier to handle things like pagination in future
changes. This adds a new dependency to the pwclient. However this is a well
known python library, and is even mentioned in the patchwork API
documentation itself.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>