api: Support msgid and archive fields in 'pwclient list'
I regularly use commands like pwclient list -m "msgid" -f'%{id}' to find
the patchwork id for a particular patch. According to the help this is
supported, but the -m commandline argument is silently ignored in the
REST API. Add support for this so the list command continues to work.
Similarly, add support for the archive argument '-a'.

Signed-off-by: Stephen Boyd <sboyd@kernel.org>