api/rest: Handle patchwork pagination
If there are many patches (or other objects) to be listed, patchwork
only returns the first 30 by default. In such a case continue fetching
objects using the "next" link provided in the Link: header.

Note this changes REST._list() to return a generator instead of a list.
All callers only use it to generate a list, so this isn't a problem.