api/rest: Add support for the submitter filter
Similar to commit 70f96f971088 ("support the delegate filter for REST
API") which added the delegate filter this implements an exact match
only and not (as promised by `pwclient list -h`) as substring match.

But for the same reason as for the delegate filter this is better than
ignoring a passed submitter.