api/rest: Let p{roject,eople,atches}_list return a generator
This has the advantage that with pagination the first entries are already
provided before later entries are queried from the server. Also if the
generator isn't walked through completely, only the needed queries are
actually executed.

All callers of these functions can cope with the returned object being a
generator now instead of a list as they just iterate once over it.